Refresh dependencies

This calls proj_update_deps(), then proj_install_deps(); use to refresh your package dependencies.

proj_refresh_deps(
  path = usethis::proj_get(),
  remove_extra = FALSE,
  dependencies = TRUE,
  ...
)

Arguments

path

character, path to the project directory. If your current working-directory in is in the project, the default will do the right thing.

remove_extra

logical, indicates to remove dependency-declarations that renv::dependencies() can't find being used.

dependencies

Which dependencies do you want to check? Can be a character vector (selecting from "Depends", "Imports", "LinkingTo", "Suggests", or "Enhances"), or a logical vector.

TRUE is shorthand for "Depends", "Imports", "LinkingTo" and "Suggests". NA is shorthand for "Depends", "Imports" and "LinkingTo" and is the default. FALSE is shorthand for no dependencies (i.e. just check this package, not its dependencies).

The value "soft" means the same as TRUE, "hard" means the same as NA.

You can also specify dependencies from one or more additional fields, common ones include:

Config/Needs/website - for dependencies used in building the pkgdown site.

Config/Needs/coverage for dependencies used in calculating test coverage.

...

other arguments passed to remotes::install_deps().

Value

Invisible NULL, called for side effects.

Examples

# not run because it produces side effects
if (FALSE) {
  proj_refresh_deps()
}