Lightweight R project management

The goal of this package is to provide some lightweight package-dependency management tools for R projects, relying on the DESCRIPTION file.

Details

The DESCRIPTION file is used, perhaps more famously, to manage dependencies for packages. Here, it is subject to the same limitation: the project will run using the latest version of all its dependencies. This assumption makes this a lightweight solution; the cost is making sure the project remain current with its dependencies.

This package's functions are mostly wrappers to usethis and renv functions. The main functions are:

proj_create()

Create a new project.

proj_use_description()

Add a DESCRIPTION file.

proj_update_deps()

Update the package-dependency declaration in DESCRIPTION.

proj_install_deps()

Install the packaage dependencies.

proj_use_github_action()

Use a GitHub Actions template.